+++ Gruntfile.js
@@ -19,6 +19,7 @@ module.exports = function(grunt) {
               './src/js/BrowserUtils.js',
               './src/js/Constants.js',
               './src/js/Color.js',
+              './src/js/DOMUtils.js',
               './src/js/AccessibilityUtils.js',
               './src/js/Properties.js',
               './src/js/AuditRule.js',