+++ src/js/AccessibilityUtils.js
@@ -16,6 +16,8 @@ goog.require('axs.browserUtils');
 goog.require('axs.color');
 goog.require('axs.color.Color');
 goog.require('axs.constants');
+goog.require('axs.dom');
+
 goog.provide('axs.utils');
 
 /**
@@ -46,60 +48,6 @@ axs.utils.LABELABLE_ELEMENTS_SELECTOR =
     'select,' +
     'textarea';
 
-/**
- * Returns the nearest ancestor which is an Element.
- * @param {Node} node
- * @return {Element}
- */
-axs.utils.parentElement = function(node) {
-    if (!node)
-        return null;
-    if (node.nodeType == Node.DOCUMENT_FRAGMENT_NODE)
-        return node.host;
-
-    var parentElement = node.parentElement;
-    if (parentElement)
-        return parentElement;
-
-    var parentNode = node.parentNode;
-    if (!parentNode)
-        return null;
-
-    switch (parentNode.nodeType) {
-    case Node.ELEMENT_NODE:
-        return /** @type {Element} */ (parentNode);
-    case Node.DOCUMENT_FRAGMENT_NODE:
-        return parentNode.host;
-    default:
-        return null;
-    }
-};
-
-/**
- * Return the corresponding element for the given node.
- * @param {Node} node
- * @return {Element}
- * @suppress {checkTypes}
- */
-axs.utils.asElement = function(node) {
-    /** @type {Element} */ var element;
-    switch (node.nodeType) {
-    case Node.COMMENT_NODE:
-        return null;  // Skip comments
-    case Node.ELEMENT_NODE:
-        element = /** (@type {Element}) */ node;
-        if (element.tagName.toLowerCase() == 'script')
-            return null;  // Skip script elements
-        break;
-    case Node.TEXT_NODE:
-        element = axs.utils.parentElement(node);
-        break;
-    default:
-        console.warn('Unhandled node type: ', node.nodeType);
-        return null;
-    }
-    return element;
-};
 
 /**
  * @param {Element} element
@@ -127,7 +75,7 @@ axs.utils.elementHasZeroArea = function(element) {
  * @return {boolean}
  */
 axs.utils.elementIsOutsideScrollArea = function(element) {
-    var parent = axs.utils.parentElement(element);
+    var parent = axs.dom.parentElement(element);
 
     var defaultView = element.ownerDocument.defaultView;
     while (parent != defaultView.document.body) {
@@ -137,7 +85,7 @@ axs.utils.elementIsOutsideScrollArea = function(element) {
         if (axs.utils.canScrollTo(element, parent) && !axs.utils.elementIsOutsideScrollArea(parent))
             return false;
 
-        parent = axs.utils.parentElement(parent);
+        parent = axs.dom.parentElement(parent);
     }
 
     return !axs.utils.canScrollTo(element, defaultView.document.body);
@@ -224,7 +172,8 @@ axs.utils.isAncestor = function(ancestor, node) {
     if (node === ancestor)
         return true;
 
-    return axs.utils.isAncestor(ancestor, node.parentNode);
+    var parentNode = axs.dom.composedParentNode(node);
+    return axs.utils.isAncestor(ancestor, parentNode);
 };
 
 /**
@@ -401,7 +350,7 @@ axs.utils.getParentBgColor = function(element) {
     /** @type {Element} */ var parent = element;
     var bgStack = [];
     var foundSolidColor = null;
-    while ((parent = axs.utils.parentElement(parent))) {
+    while ((parent = axs.dom.parentElement(parent))) {
         var computedStyle = window.getComputedStyle(parent, null);
         if (!computedStyle)
             continue;
@@ -564,14 +513,14 @@ axs.utils.hasLabel = function(element) {
             return true;
     }
 
-    var parent = axs.utils.parentElement(element);
+    var parent = axs.dom.parentElement(element);
     while (parent) {
         if (parent.tagName.toLowerCase() == 'label') {
             var parentLabel = /** HTMLLabelElement */ parent;
             if (parentLabel.control == element)
                 return true;
         }
-        parent = axs.utils.parentElement(parent);
+        parent = axs.dom.parentElement(parent);
     }
     return false;
 };
@@ -605,7 +554,7 @@ axs.utils.isElementDisabled = function(element) {
             axs.browserUtils.matchSelector(element, 'fieldset>legend:first-of-type *')) {
         return false;
     }
-    for (var next = element; next !== null; next = axs.utils.parentElement(next)) {
+    for (var next = element; next !== null; next = axs.dom.parentElement(next)) {
         if(axs.utils.isNativelyDisableable(next) && next.hasAttribute('disabled')) {
             return true;
         }
@@ -645,8 +594,8 @@ axs.utils.isElementOrAncestorHidden = function(element) {
     if (axs.utils.isElementHidden(element))
         return true;
 
-    if (axs.utils.parentElement(element))
-        return axs.utils.isElementOrAncestorHidden(axs.utils.parentElement(element));
+    if (axs.dom.parentElement(element))
+        return axs.utils.isElementOrAncestorHidden(axs.dom.parentElement(element));
     else
         return false;
 };